requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
pydantic[email]
python-dotenv>=0.19.0
openai>=1.0.0
pytest>=6.0.0
httpx>=0.24.0 